requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==4.4.2
idna==3.3
imageio==2.16.1
imageio-ffmpeg==0.4.5
instagrapi==1.16.14
moviepy==1.0.3
numpy==1.21.5
Pillow==9.0.1
praw==7.6.0
prawcore==2.3.0
proglog==0.1.9
pycryptodomex==3.12.0
pydantic==1.9.0
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.27.1
tqdm==4.63.0
typing-extensions==4.1.1
update-checker==0.18.0
urllib3==1.26.8
websocket-client==1.3.1
